test: rename e2e/util to e2e/e2eutil
