+++ test/e2e/containerd_test.go
@@ -19,7 +19,7 @@ package e2e
 import (
 	"fmt"
 
-	"d7y.io/dragonfly/v2/test/e2e/util"
+	"d7y.io/dragonfly/v2/test/e2e/e2eutil"
 	. "github.com/onsi/ginkgo" //nolint
 	. "github.com/onsi/gomega" //nolint
 )
@@ -27,15 +27,15 @@ import (
 var _ = Describe("Containerd with CRI support", func() {
 	Context("docker.io/library/busybox:latest image", func() {
 		It("pull should be ok", func() {
-			out, err := util.CrictlCommand("pull", "d7y.io/library/busybox:latest").CombinedOutput()
-			Expect(err).NotTo(HaveOccurred())
+			out, err := e2eutil.CriCtlCommand("pull", "d7y.io/library/busybox:latest").CombinedOutput()
 			fmt.Println(string(out))
+			Expect(err).NotTo(HaveOccurred())
 		})
 
 		It("rmi should be ok", func() {
-			out, err := util.CrictlCommand("rmi", "d7y.io/library/busybox:latest").CombinedOutput()
-			Expect(err).NotTo(HaveOccurred())
+			out, err := e2eutil.CriCtlCommand("rmi", "d7y.io/library/busybox:latest").CombinedOutput()
 			fmt.Println(string(out))
+			Expect(err).NotTo(HaveOccurred())
 		})
 	})
 })
