+++ test/e2e/e2eutil/exec.go
@@ -1,4 +1,4 @@
-package util
+package e2eutil
 
 import "os/exec"
 
@@ -8,8 +8,12 @@ func DockerCommand(arg ...string) *exec.Cmd {
 	return exec.Command("docker", extArgs...)
 }
 
-func CrictlCommand(arg ...string) *exec.Cmd {
+func CriCtlCommand(arg ...string) *exec.Cmd {
 	extArgs := []string{"/usr/local/bin/crictl"}
 	extArgs = append(extArgs, arg...)
 	return DockerCommand(extArgs...)
+}
+
+func KubeCtlCommand(arg ...string) *exec.Cmd {
+	return exec.Command("kubectl", arg...)
 }
